feat: finish split top card
