+++ src/components/Cards/EarnCard/EarnCard.styles.ts
@@ -101,16 +101,32 @@ export const ListItemEarnCardTagContainer = styled(Stack)(({ theme }) => ({
 
 export const TopEarnCardContainer = styled(EarnCardContainer)(({ theme }) => ({
   padding: theme.spacing(3),
+  gap: theme.spacing(1),
+  minHeight: 312,
+  backgroundColor: (theme.vars || theme).palette.surface2.main,
+  display: 'flex',
+  flexDirection: 'column',
 }));
 
 export const TopEarnCardHeaderContainer = styled(Stack)(({ theme }) => ({
   gap: theme.spacing(2),
 }));
 
-export const TopEarnCardContentContainer = styled(Stack)(({ theme }) => ({
+interface TopEarnCardContentContainerProps {
+  isMain?: boolean;
+}
+
+export const TopEarnCardContentContainer = styled(Stack, {
+  shouldForwardProp: (prop) => prop !== 'isMain',
+})<TopEarnCardContentContainerProps>(({ theme, isMain }) => ({
   gap: theme.spacing(2),
+  '& p': {
+    ...(isMain ? theme.typography.h3 : theme.typography.h4),
+    margin: 0,
+  },
 }));
 
 export const TopEarnCardFooterContainer = styled(Stack)(({ theme }) => ({
   gap: theme.spacing(2),
+  marginTop: 'auto',
 }));
