+++ src/components/Cards/EarnCard/variants/TopHiglight.tsx
@@ -7,16 +7,13 @@ interface Props {
 
 const HighlightedSpan = styled('span')(({ theme }) => ({
   color: (theme.vars || theme).palette.accent1Alt.main,
-  fontWeight: theme.typography.fontWeightMedium,
+  fontWeight: 'inherit',
   ...theme.applyStyles('light', {
     color: (theme.vars || theme).palette.accent1.main,
   }),
 }));
 
-export const TopHighlight: FC<PropsWithChildren<Props>> = ({
-  type,
-  children,
-}) => {
+export const TopHighlight: FC<PropsWithChildren<Props>> = ({ children }) => {
   // TODO: LF-14990: Generate clickable links / filters events if possible.
   return <HighlightedSpan>{children}</HighlightedSpan>;
 };
